+++ Sources/Controllers/MessagesViewController.swift
@@ -88,12 +88,14 @@ open class MessagesViewController: UIViewController {
     }
 
     open override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
         // Hack to prevent animation of the contentInset after viewDidAppear
         if isFirstLayout {
             defer { isFirstLayout = false }
             addKeyboardObservers()
             messageCollectionViewBottomInset = keyboardOffsetFrame.height
         }
+        adjustScrollViewInset()
     }
 
     // MARK: - Initializers
@@ -143,7 +145,6 @@ open class MessagesViewController: UIViewController {
             let trailing = messagesCollectionView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
             NSLayoutConstraint.activate([top, bottom, trailing, leading])
         }
-        adjustScrollViewInset()
     }
     
     @objc
@@ -151,7 +152,10 @@ open class MessagesViewController: UIViewController {
         if #available(iOS 11.0, *) {
             // No need to add to the top contentInset
         } else {
-            let navigationBarInset = navigationController?.navigationBar.frame.height ?? 0
+            var navigationBarInset: CGFloat = 0
+            if let navigationController = navigationController, !navigationController.isNavigationBarHidden {
+                navigationBarInset = navigationController.navigationBar.frame.height
+            }
             let statusBarInset: CGFloat = UIApplication.shared.isStatusBarHidden ? 0 : 20
             let topInset = navigationBarInset + statusBarInset
             messagesCollectionView.contentInset.top = topInset